+++ app/src/main/java/com/example/wayoflife/workouts/RunningActivity.java
@@ -173,7 +173,7 @@ public class RunningActivity extends AppCompatActivity {
                 Constants.PROFILE_INFO_FILENAME,
                 Context.MODE_PRIVATE);
 
-        int peso = Integer.parseInt(sharedPref.getString(
+        double peso = Double.parseDouble(sharedPref.getString(
                 Constants.PESO, "0"));
 
         switch(attivitaDiProvenienza){
